requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
facebook2==2.2.1
idna==2.8
requests==2.22.0
urllib3==1.25.3
